feat: deprecate gomnd, add mnd
