+++ docs/src/docs/usage/false-positives.mdx
@@ -49,8 +49,8 @@ In the following example, all the reports from the linters (`linters`) that cont
 issues:
   exclude-rules:
     - linters:
-        - gomnd
-      text: "mnd: Magic number: 9"
+        - mnd
+      text: "Magic number: 9"
 ```
 
 In the following example, all the reports from the linters (`linters`) that originated from the source (`source`) are excluded:
